+++ src/types/tbtc.ts
@@ -1,7 +1,7 @@
 import { UnspentTransactionOutput } from "@keep-network/tbtc-v2.ts/dist/src/bitcoin"
 import { UpdateStateActionPayload } from "./state"
 import { FetchingState } from "."
-import { BridgeActivity, BridgeProcess } from "../threshold-ts/types/tbtc"
+import { BridgeActivity, BridgeProcess } from "../threshold-ts/tbtc"
 export type { UnspentTransactionOutputPlainObject } from "../threshold-ts/types"
 
 export interface TbtcState {
